Updated gsap package to business version
The gsap package has been updated from the standard version to the business version. This change includes updates in .npmrc, package.json, and pnpm-lock.yaml files. The new configuration for authentication and registry is added in .npmrc file. In package.json and pnpm-lock.yaml files, the specifier and version of gsap have been changed to reflect this update.